Rename output/results-doc to repl-window/repl-doc
