+++ src/lsp/client/client.ts
@@ -1,7 +1,7 @@
 import * as messages from './messages';
 import { provideSignatureHelp } from '../../providers/signature';
 import { provideHover } from '../../providers/hover';
-import { isResultsDoc } from '../../results-output/results-doc';
+import { isResultsDoc } from '../../repl-window/repl-doc';
 import * as vscode_lsp from 'vscode-languageclient/node';
 import * as defs from '../definitions';
 import * as config from '../../config';
