+++ src/nrepl/jack-in.ts
@@ -13,7 +13,7 @@ import {
   getConnectSequences,
 } from './connectSequence';
 import * as projectTypes from './project-types';
-import * as outputWindow from '../results-output/results-doc';
+import * as outputWindow from '../repl-window/repl-doc';
 import { JackInTerminal, JackInTerminalOptions, createCommandLine } from './jack-in-terminal';
 import * as liveShareSupport from '../live-share';
 import { getConfig } from '../config';
