+++ src/nrepl/repl-session.ts
@@ -1,6 +1,6 @@
 import { NReplSession } from '.';
 import { cljsLib, tryToGetDocument, getFileType } from '../utilities';
-import * as outputWindow from '../results-output/results-doc';
+import * as outputWindow from '../repl-window/repl-doc';
 import { isUndefined } from 'lodash';
 
 function getSession(fileType?: string): NReplSession {
