+++ src/results-output/util.ts
@@ -1,5 +1,5 @@
 import { takeWhile } from 'lodash';
-import type { ResultsBuffer } from './results-doc';
+import type { ResultsBuffer } from '../repl-window/repl-doc';
 
 function addToHistory(history: string[], content?: string): string[] {
   if (content) {
